# feat: reuse seenEdge and seenNode to reduce garbage collection pressure