@kalyanac 2020-10-23
Thank you for the PR. In the PR descriptio, can you add the timing you measured with this PR and the current code?